"Fix" sagepay and paypal
The sagepay code wasn't working right, the adminhtml code wasn't setting
the `config_path` so it was being set to `sagepaysuite/method/filter`
but when the filter was asking for the payment method's config data it
was looking for `payment/method/fitler`. Magento also moved the paypal
methods at some point, so the code to use them broke as well.

This patch "fixes" both, and break all existing installations. Instead
of creating a config key under the payment method / carrier it creates a
key in n98_checkoutfilters/filter/code. For paypal I gave up trying to
get that deep into the xml, and it loads one set of filters in the top
paypal settings box which it applies to all the different paypal
methods.